Added MVC architecture for JS. Added API fetching
Added API fetching of some postman API. For now just a simple fetch with the first element from the data array of the API. Next to add is fetching the specific opportunity/job with the slug property. Also need to add views to render and a new html page for that I think.