Added a 5th smoke test
It works, the previous one has an difficult bug. We will probably change the 4th smoke test case because of the bug. Bug is about the apply form not opening and being visible. When we directly call the toggleWindow method on it, it opens and becomes visible. However when a click function or dispachEvent is used on the apply now button it does not open.